Move modules to core folder
