+++ src/neural_data_simulator/core/encoder.py
@@ -3,10 +3,10 @@ import contextlib
 import logging
 from typing import Iterator, Optional, Protocol, runtime_checkable
 
-from neural_data_simulator.inputs import Input
-from neural_data_simulator.models import EncoderModel
-from neural_data_simulator.outputs import Output
-from neural_data_simulator.samples import Samples
+from neural_data_simulator.core.inputs import Input
+from neural_data_simulator.core.models import EncoderModel
+from neural_data_simulator.core.outputs import Output
+from neural_data_simulator.core.samples import Samples
 from neural_data_simulator.util.runtime import open_connection
