+++ src/neural_data_simulator/core/timing.py
