+++ src/neural_data_simulator/scripts/run_encoder.py
@@ -16,18 +16,18 @@ from typing import Callable, cast, Optional, Union
 
 import numpy as np
 
-from neural_data_simulator import encoder
-from neural_data_simulator import inputs
-from neural_data_simulator import models
-from neural_data_simulator import outputs
-from neural_data_simulator import runner
-from neural_data_simulator import timing
-from neural_data_simulator.outputs import LSLOutputDevice
-from neural_data_simulator.samples import Samples
+from neural_data_simulator.core import encoder
+from neural_data_simulator.core import inputs
+from neural_data_simulator.core import models
+from neural_data_simulator.core import outputs
+from neural_data_simulator.core import runner
+from neural_data_simulator.core import timing
+from neural_data_simulator.core.outputs import LSLOutputDevice
+from neural_data_simulator.core.samples import Samples
 from neural_data_simulator.scripts.errors import InvalidPluginError
-from neural_data_simulator.settings import EncoderEndpointType
-from neural_data_simulator.settings import EncoderSettings
-from neural_data_simulator.settings import Settings
+from neural_data_simulator.core.settings import EncoderEndpointType
+from neural_data_simulator.core.settings import EncoderSettings
+from neural_data_simulator.core.settings import Settings
 from neural_data_simulator.util.runtime import configure_logger
 from neural_data_simulator.util.runtime import get_abs_path
 from neural_data_simulator.util.runtime import initialize_logger
