+++ src/neural_data_simulator/scripts/run_ephys_generator.py
@@ -17,21 +17,21 @@ from typing import cast, Optional
 
 import numpy as np
 
-from neural_data_simulator import inputs
-from neural_data_simulator import outputs
-from neural_data_simulator.ephys_generator import ContinuousData
-from neural_data_simulator.ephys_generator import LSLSpikeRateInputAdapter
-from neural_data_simulator.ephys_generator import NoiseData
-from neural_data_simulator.ephys_generator import ProcessOutput
-from neural_data_simulator.ephys_generator import SpikeRateInput
-from neural_data_simulator.ephys_generator import SpikeRateTestingInput
-from neural_data_simulator.ephys_generator import Spikes
-from neural_data_simulator.ephys_generator import Waveforms
-from neural_data_simulator.health_checker import HealthChecker
-from neural_data_simulator.outputs import StreamConfig
-from neural_data_simulator.settings import EphysGeneratorEndpointType
-from neural_data_simulator.settings import EphysGeneratorSettings
-from neural_data_simulator.settings import Settings
+from neural_data_simulator.core import inputs
+from neural_data_simulator.core import outputs
+from neural_data_simulator.core.ephys_generator import ContinuousData
+from neural_data_simulator.core.ephys_generator import LSLSpikeRateInputAdapter
+from neural_data_simulator.core.ephys_generator import NoiseData
+from neural_data_simulator.core.ephys_generator import ProcessOutput
+from neural_data_simulator.core.ephys_generator import SpikeRateInput
+from neural_data_simulator.core.ephys_generator import SpikeRateTestingInput
+from neural_data_simulator.core.ephys_generator import Spikes
+from neural_data_simulator.core.ephys_generator import Waveforms
+from neural_data_simulator.core.health_checker import HealthChecker
+from neural_data_simulator.core.outputs import StreamConfig
+from neural_data_simulator.core.settings import EphysGeneratorEndpointType
+from neural_data_simulator.core.settings import EphysGeneratorSettings
+from neural_data_simulator.core.settings import Settings
 from neural_data_simulator.util.runtime import configure_logger
 from neural_data_simulator.util.runtime import initialize_logger
 from neural_data_simulator.util.runtime import unwrap
