+++ src/neural_data_simulator/util/runtime.py
@@ -10,9 +10,9 @@ import sys
 from types import ModuleType
 from typing import Any, Optional, Union
 
-from neural_data_simulator.inputs import Input
-from neural_data_simulator.outputs import Output
-from neural_data_simulator.settings import LogLevel
+from neural_data_simulator.core.inputs import Input
+from neural_data_simulator.core.outputs import Output
+from neural_data_simulator.core.settings import LogLevel
 
 logger = logging.getLogger(__name__)
