+++ src/plugins/examples/gamepad_preprocessor.py
@@ -1,6 +1,6 @@
 """Preprocessor for LSL Gamepad app for velocity tuning model."""
-from neural_data_simulator.encoder import Processor
-from neural_data_simulator.samples import Samples
+from neural_data_simulator.core.encoder import Processor
+from neural_data_simulator.core.samples import Samples
 
 
 class GamepadPreprocessor(Processor):
