+++ src/tasks/center_out_reach/task_runner.py
@@ -20,11 +20,11 @@ from tasks.center_out_reach.metrics import MetricsCollector
 from tasks.center_out_reach.task_state import TaskState
 from tasks.center_out_reach.task_window import TaskWindow
 
-from neural_data_simulator import inputs
-from neural_data_simulator import outputs
-from neural_data_simulator.filters import LowpassFilter
-from neural_data_simulator.samples import Samples
-from neural_data_simulator.timing import Timer
+from neural_data_simulator.core import inputs
+from neural_data_simulator.core import outputs
+from neural_data_simulator.core.filters import LowpassFilter
+from neural_data_simulator.core.samples import Samples
+from neural_data_simulator.core.timing import Timer
 
 
 class VelocityScaler(Protocol):
